move karma config to separate file (:tophat: @codebling)
